added another issue to the list
added functionality for if no tiles are moved (only to up button press 
for now)

changed 2D array to be a 2D array of objects

added combined functionality to see if the tiles have already been 
combined or not. used for more than 2 of the same number on the same row 
or column being combined to only move once instead of combining all at 
once. (only implemented on up button press)

added a function to clear tile moved flag and combined flag for each 
button press